Update Gemfile Dependency
‘pg’ wasn’t set to a particular version and Bundler was failing to
`bundle install` so I’ve decided to make the dependency use the current
version I’ve installed on my machine. Bundler was successful and I can
now run my app locally.

This is mainly a reminder to myself because I’ve run into this problem
yesterday and nearly lost my mind trying to fix it, as I do with Rails
90% of the time.